feat: renamed all react components file names to PascalCase
